refactor: improve drag layer mechanism and add offset safety check
