+++ packages/react-docs/pages/components/tree-view/dnd/ConditionalWrapper.js
@@ -1,0 +1,9 @@
+const ConditionalWrapper = ({
+  children,
+  condition,
+  wrapper,
+}) => {
+  return condition ? wrapper(children) : children;
+};
+
+export default ConditionalWrapper;
