+++ packages/react-docs/pages/components/tree-view/dnd/TableView.js
@@ -7,20 +7,11 @@ import {
 } from '@tanstack/react-table';
 import {
   Box,
-  Button,
-  ButtonGroup,
   Checkbox,
-  Divider,
   Flex,
   Icon,
   Image,
-  Input,
-  Menu,
-  MenuButton,
-  MenuList,
-  MenuItem,
   OverflowTooltip,
-  Space,
   Table,
   TableHeader,
   TableHeaderRow,
@@ -29,7 +20,6 @@ import {
   TableRow,
   TableCell,
   TableScrollbar,
-  Text,
   useColorMode,
   useColorStyle,
   useTheme,
@@ -39,39 +29,12 @@ import {
 } from '@tonic-ui/utils';
 import { ensureArray, ensureNumber } from 'ensure-type';
 import _ from 'lodash';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { useDrag } from 'react-dnd';
-import { getEmptyImage } from 'react-dnd-html5-backend'
+import React, { useEffect, useMemo, useState } from 'react';
 import AutoSizer from 'react-virtualized-auto-sizer';
-import iconDrag from './icons/icon-drag.svg';
-
-const Draggable = ({
-  canDrag: canDragProp,
-  children,
-  item: itemProp,
-}) => {
-  const [collectedProps, dragRef, dragPreviewRef] = useDrag({
-    type: 'dnd',
-    item: itemProp,
-    canDrag: canDragProp,
-    collect: (monitor) => {
-      // DragSourceMonitor
-      // https://react-dnd.github.io/react-dnd/docs/api/drag-source-monitor
-      return {
-        isDragging: monitor.isDragging(),
-      };
-    },
-  });
-
-  useEffect(() => {
-    dragPreviewRef(getEmptyImage(), { captureDraggingState: true });
-  }, [dragPreviewRef]);
-
-  return children({
-    dragRef,
-    isDragging: collectedProps.isDragging,
-  });
-};
+import ConditionalWrapper from './ConditionalWrapper';
+import Draggable from './Draggable';
+import TablePagination from './TablePagination';
+import iconDrag from './icon-drag.svg';
 
 /**
  * Uses canvas.measureText to compute and return the width of the given text of given font in pixels.
@@ -90,7 +53,7 @@ const getTextWidth = (text, font) => {
   return metrics.width || 0;
 };
 
-const DndTable = ({
+const Component = ({
   node,
   data,
   ...rest
@@ -544,168 +507,4 @@ const DndTable = ({
   );
 };
 
-const TablePagination = ({
-  count,
-  defaultPage = 1,
-  defaultRowsPerPage: defaultRowsPerPageProp,
-  onPageChange: onPageChangeProp,
-  onRowsPerPageChange: onRowsPerPageChangeProp,
-  page: pageProp,
-  rowsPerPage: rowsPerPageProp,
-  rowsPerPageOptions = [10, 25, 50, 100],
-  showFirstButton = false,
-  showLastButton = false,
-}) => {
-  const [colorMode] = useColorMode();
-  const [colorStyle] = useColorStyle({ colorMode });
-  const defaultRowsPerPage = defaultRowsPerPageProp ?? ensureArray(rowsPerPageOptions)[0];
-  const [page, setPage] = useState(pageProp ?? defaultPage);
-  const [rowsPerPage, setRowsPerPage] = useState(rowsPerPageProp ?? defaultRowsPerPage);
-
-  useEffect(() => {
-    const isControlled = (pageProp !== undefined);
-    if (isControlled) {
-      setPage(pageProp);
-    }
-  }, [pageProp]);
-
-  useEffect(() => {
-    const isControlled = (rowsPerPageProp !== undefined);
-    if (isControlled) {
-      setRowsPerPage(rowsPerPageProp);
-    }
-  }, [rowsPerPageProp]);
-
-  const onPageChange = useCallback((nextPage) => {
-    const isControlled = (pageProp !== undefined);
-    if (!isControlled) {
-      setPage(nextPage);
-    }
-
-    if (typeof onPageChangeProp === 'function') {
-      onPageChangeProp(nextPage);
-    }
-  }, [pageProp, onPageChangeProp]);
-
-  const onRowsPerPageChange = useCallback((nextRowsPerPage) => {
-    const isControlled = (rowsPerPageProp !== undefined);
-    if (!isControlled) {
-      setRowsPerPage(nextRowsPerPage);
-    }
-
-    if (typeof onRowsPerPageChangeProp === 'function') {
-      onRowsPerPageChangeProp(nextRowsPerPage);
-    }
-  }, [rowsPerPageProp, onRowsPerPageChangeProp]);
-
-  const totalPages = Math.ceil(count / rowsPerPage);
-  const handlePageChange = (event) => {
-    const nextPage = ensureNumber(event.target.value);
-    if (nextPage <= 1) {
-      onPageChange(1);
-    } else if (nextPage >= totalPages) {
-      onPageChange(totalPages);
-    } else {
-      onPageChange(nextPage);
-    }
-  };
-  const handleRowsPerPageChange = (event) => {
-    const nextRowsPerPage = ensureNumber(event.target.value);
-    if (nextRowsPerPage > 0) {
-      onPageChange(1);
-      onRowsPerPageChange(nextRowsPerPage);
-    }
-  };
-  const canPreviousPage = (page > 1);
-  const canNextPage = (page < totalPages);
-
-  return (
-    <Flex
-      alignItems="center"
-      justifyContent="flex-end"
-      backgroundColor={colorStyle.background.secondary}
-      px="6x"
-      py="3x"
-    >
-      <Text mr="2x">
-        Total: {count}
-      </Text>
-      <Divider
-        orientation="vertical"
-        height="6x"
-      />
-      <Menu>
-        <MenuButton variant="ghost">
-          {rowsPerPage} per page
-        </MenuButton>
-        <MenuList
-          onClick={handleRowsPerPageChange}
-          width="100%"
-        >
-          {rowsPerPageOptions.map((option) => (
-            <MenuItem
-              key={option}
-              value={option}
-            >
-              {option}
-            </MenuItem>
-          ))}
-        </MenuList>
-      </Menu>
-      <Divider
-        orientation="vertical"
-        height="6x"
-      />
-      <Space width="2x" />
-      <Input
-        width="10x"
-        px={0}
-        textAlign="center"
-        onChange={handlePageChange}
-        value={page}
-      />
-      <Space width="2x" />
-      <Text>/</Text>
-      <Space width="2x" />
-      <Text>{totalPages}</Text>
-      <Space width="2x" />
-      <ButtonGroup
-        variant="secondary"
-        sx={{
-          '> *:not(:first-of-type)': {
-            marginLeft: -1
-          }
-        }}
-      >
-        <Button
-          width="8x"
-          disabled={!canPreviousPage}
-          onClick={(event) => {
-            onPageChange(page - 1);
-          }}
-        >
-          <Icon icon="angle-left" />
-        </Button>
-        <Button
-          width="8x"
-          disabled={!canNextPage}
-          onClick={(event) => {
-            onPageChange(page + 1);
-          }}
-        >
-          <Icon icon="angle-right" />
-        </Button>
-      </ButtonGroup>
-    </Flex>
-  );
-};
-
-const ConditionalWrapper = ({
-  children,
-  condition,
-  wrapper,
-}) => {
-  return condition ? wrapper(children) : children;
-};
-
-export default DndTable;
+export default Component;
