+++ packages/react-docs/pages/components/tree-view/dnd/useRefresh.js
@@ -1,0 +1,10 @@
+import { useCallback, useState } from 'react';
+
+const useRefresh = () => {
+  const [, rerender] = useState();
+  return useCallback(() => {
+    rerender({});
+  }, []);
+};
+
+export default useRefresh;
